typed_generational_arena
Trait GenerationalIndex

pub trait GenerationalIndex: FixedGenerationalIndex {
    // Required method
    fn increment_generation(&mut self);
}

A type which can be used as the index of a generation, which can be incremented

Required Methods

fn increment_generation(&mut self)

Increment the generation of this object. May wrap or panic on overflow depending on type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.